separated solution to solution (red node) and solutionstate (green node)
I didn't make workspace to hold onto solution state though. but we can if we want do.
