+++ src/Workspaces/CSharpTest/CodeGeneration/AddImportsTests.cs
@@ -485,10 +485,10 @@ class C
         [WorkItem(9228, "https://github.com/dotnet/roslyn/issues/9228")]
         public async Task TestDoNotAddDuplicateImportIfNamespaceIsDefinedInSourceAndExternalAssembly()
         {
-            var externalCode = 
+            var externalCode =
 @"namespace N.M { public class A : System.Attribute { } }";
 
-            var code = 
+            var code =
 @"using System;
 using N.M;
 
@@ -510,7 +510,7 @@ class C
 
             var options = document.Project.Solution.Workspace.Options;
 
-            var compilation = await document.Project.Solution.GetCompilationAsync(document.Project, CancellationToken.None);
+            var compilation = await document.Project.GetCompilationAsync(CancellationToken.None);
             ImmutableArray<Diagnostic> compilerDiagnostics = compilation.GetDiagnostics(CancellationToken.None);
             Assert.Empty(compilerDiagnostics.Where(d => d.Severity == DiagnosticSeverity.Error));
