+++ src/Workspaces/Core/Portable/Workspace/Solution/ProjectState.cs
@@ -188,47 +188,6 @@ namespace Microsoft.CodeAnalysis
             return doc;
         }
 
-        public ProjectId Id
-        {
-            get { return this.ProjectInfo.Id; }
-        }
-
-        public string FilePath
-        {
-            get { return this.ProjectInfo.FilePath; }
-        }
-
-        public string OutputFilePath
-        {
-            get { return this.ProjectInfo.OutputFilePath; }
-        }
-
-        public HostLanguageServices LanguageServices
-        {
-            get { return _languageServices; }
-        }
-
-        public string Name
-        {
-            get { return this.ProjectInfo.Name; }
-        }
-
-        public bool IsSubmission
-        {
-            get { return this.ProjectInfo.IsSubmission; }
-        }
-
-        public Type HostObjectType
-        {
-            get { return this.ProjectInfo.HostObjectType; }
-        }
-
-        [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public VersionStamp Version
-        {
-            get { return this.ProjectInfo.Version; }
-        }
-
         public AnalyzerOptions AnalyzerOptions
         {
             get
@@ -264,88 +223,73 @@ namespace Microsoft.CodeAnalysis
         }
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public ProjectInfo ProjectInfo
-        {
-            get { return _projectInfo; }
-        }
+        public ProjectId Id => this.ProjectInfo.Id;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public string AssemblyName
-        {
-            get { return this.ProjectInfo.AssemblyName; }
-        }
+        public string FilePath => this.ProjectInfo.FilePath;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public CompilationOptions CompilationOptions
-        {
-            get { return this.ProjectInfo.CompilationOptions; }
-        }
+        public string OutputFilePath => this.ProjectInfo.OutputFilePath;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public ParseOptions ParseOptions
-        {
-            get { return this.ProjectInfo.ParseOptions; }
-        }
+        public HostLanguageServices LanguageServices => _languageServices;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public IReadOnlyList<MetadataReference> MetadataReferences
-        {
-            get { return this.ProjectInfo.MetadataReferences; }
-        }
+        public string Name => this.ProjectInfo.Name;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public IReadOnlyList<AnalyzerReference> AnalyzerReferences
-        {
-            get { return this.ProjectInfo.AnalyzerReferences; }
-        }
+        public bool IsSubmission => this.ProjectInfo.IsSubmission;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public IReadOnlyList<ProjectReference> ProjectReferences
-        {
-            get { return this.ProjectInfo.ProjectReferences; }
-        }
+        public Type HostObjectType => this.ProjectInfo.HostObjectType;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public bool HasAllInformation
-        {
-            get { return this.ProjectInfo.HasAllInformation; }
-        }
+        public bool SupportsCompilation => this.LanguageServices.GetService<ICompilationFactoryService>() != null;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public bool HasDocuments
-        {
-            get { return _documentIds.Count > 0; }
-        }
+        public VersionStamp Version => this.ProjectInfo.Version;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public IEnumerable<DocumentState> OrderedDocumentStates
-        {
-            get { return this.DocumentIds.Select(GetDocumentState); }
-        }
+        public ProjectInfo ProjectInfo => _projectInfo;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public IReadOnlyList<DocumentId> DocumentIds
-        {
-            get { return _documentIds; }
-        }
+        public string AssemblyName => this.ProjectInfo.AssemblyName;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        public IReadOnlyList<DocumentId> AdditionalDocumentIds
-        {
-            get { return _additionalDocumentIds; }
-        }
+        public CompilationOptions CompilationOptions => this.ProjectInfo.CompilationOptions;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        private ImmutableDictionary<DocumentId, DocumentState> DocumentStates
-        {
-            get { return _documentStates; }
-        }
+        public ParseOptions ParseOptions => this.ProjectInfo.ParseOptions;
 
         [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
-        private ImmutableDictionary<DocumentId, TextDocumentState> AdditionalDocumentStates
-        {
-            get { return _additionalDocumentStates; }
-        }
+        public IReadOnlyList<MetadataReference> MetadataReferences => this.ProjectInfo.MetadataReferences;
+
+        [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
+        public IReadOnlyList<AnalyzerReference> AnalyzerReferences => this.ProjectInfo.AnalyzerReferences;
+
+        [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
+        public IReadOnlyList<ProjectReference> ProjectReferences => this.ProjectInfo.ProjectReferences;
+
+        [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
+        public bool HasAllInformation => this.ProjectInfo.HasAllInformation;
+
+        [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
+        public bool HasDocuments => _documentIds.Count > 0;
+
+        [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
+        public IEnumerable<DocumentState> OrderedDocumentStates => this.DocumentIds.Select(GetDocumentState);
+
+        [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
+        public IReadOnlyList<DocumentId> DocumentIds => _documentIds;
+
+        [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
+        public IReadOnlyList<DocumentId> AdditionalDocumentIds => _additionalDocumentIds;
+
+        [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
+        private ImmutableDictionary<DocumentId, DocumentState> DocumentStates => _documentStates;
+
+        [DebuggerBrowsable(DebuggerBrowsableState.Collapsed)]
+        private ImmutableDictionary<DocumentId, TextDocumentState> AdditionalDocumentStates => _additionalDocumentStates;
 
         public bool ContainsDocument(DocumentId documentId)
         {
