+++ src/Workspaces/Core/Portable/Workspace/Solution/Solution.cs
@@ -4,11 +4,9 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
-using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.ErrorReporting;
 using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.Shared.Extensions;
 using Microsoft.CodeAnalysis.Text;
@@ -22,281 +20,72 @@ namespace Microsoft.CodeAnalysis
     /// </summary>
     public partial class Solution
     {
-        // branch id for this solution
-        private readonly BranchId _branchId;
-
-        // the version of the workspace this solution is from
-        private readonly int _workspaceVersion;
-
-        private readonly SolutionServices _solutionServices;
-        private readonly SolutionId _id;
-        private readonly string _filePath;
-        private readonly IReadOnlyList<ProjectId> _projectIds;
-        private readonly ImmutableDictionary<ProjectId, ProjectState> _projectIdToProjectStateMap;
-        private readonly ImmutableDictionary<string, ImmutableArray<DocumentId>> _linkedFilesMap;
-        private readonly VersionStamp _version;
-        private readonly Lazy<VersionStamp> _lazyLatestProjectVersion;
-        private readonly ProjectDependencyGraph _dependencyGraph;
+        // SolutionState that doesn't hold onto Project/Document
+        private readonly SolutionState _state;
 
         // Values for all these are created on demand.
         private ImmutableHashMap<ProjectId, Project> _projectIdToProjectMap;
-        private ImmutableDictionary<ProjectId, CompilationTracker> _projectIdToTrackerMap;
 
-        private Solution(
-            BranchId branchId,
-            int workspaceVersion,
-            SolutionServices solutionServices,
-            SolutionId id,
-            string filePath,
-            IEnumerable<ProjectId> projectIds,
-            ImmutableDictionary<ProjectId, ProjectState> idToProjectStateMap,
-            ImmutableDictionary<ProjectId, CompilationTracker> projectIdToTrackerMap,
-            ImmutableDictionary<string, ImmutableArray<DocumentId>> linkedFilesMap,
-            ProjectDependencyGraph dependencyGraph,
-            VersionStamp version,
-            Lazy<VersionStamp> lazyLatestProjectVersion)
+        private Solution(SolutionState state)
         {
-            _branchId = branchId;
-            _workspaceVersion = workspaceVersion;
-            _id = id;
-            _filePath = filePath;
-            _solutionServices = solutionServices;
-            _projectIds = projectIds.ToImmutableReadOnlyListOrEmpty();
-            _projectIdToProjectStateMap = idToProjectStateMap;
-            _projectIdToTrackerMap = projectIdToTrackerMap;
-            _linkedFilesMap = linkedFilesMap;
-            _dependencyGraph = dependencyGraph;
             _projectIdToProjectMap = ImmutableHashMap<ProjectId, Project>.Empty;
-            _version = version;
-            _lazyLatestProjectVersion = lazyLatestProjectVersion;
-
-            CheckInvariants();
+            _state = state;
         }
 
-        internal Solution(
-            Workspace workspace,
-            SolutionInfo info)
-            : this(
-                workspace.PrimaryBranchId,
-                workspaceVersion: 0,
-                solutionServices: new SolutionServices(workspace),
-                id: info.Id,
-                filePath: info.FilePath,
-                version: info.Version,
-                projectIds: null,
-                idToProjectStateMap: ImmutableDictionary<ProjectId, ProjectState>.Empty,
-                projectIdToTrackerMap: ImmutableDictionary<ProjectId, CompilationTracker>.Empty,
-                linkedFilesMap: ImmutableDictionary.Create<string, ImmutableArray<DocumentId>>(StringComparer.OrdinalIgnoreCase),
-                dependencyGraph: ProjectDependencyGraph.Empty,
-                lazyLatestProjectVersion: null)
+        internal Solution(Workspace workspace, SolutionInfo info)
+            : this(new SolutionState(workspace, info))
         {
-            _lazyLatestProjectVersion = new Lazy<VersionStamp>(() => ComputeLatestProjectVersion());
         }
 
-        internal Solution WithNewWorkspace(Workspace workspace, int workspaceVersion)
-        {
-            var services = workspace != _solutionServices.Workspace
-                ? new SolutionServices(workspace)
-                : _solutionServices;
+        internal SolutionState State => _state;
 
-            // Note: this will potentially have problems if the workspace services are different, as some services
-            // get locked-in by document states and project states when first constructed.
-            return CreatePrimarySolution(branchId: workspace.PrimaryBranchId, workspaceVersion: workspaceVersion, services: services);
-        }
+        internal int WorkspaceVersion => _state.WorkspaceVersion;
 
-        private VersionStamp ComputeLatestProjectVersion()
-        {
-            // this may produce a version that is out of sync with the actual Document versions.
-            var latestVersion = VersionStamp.Default;
-            foreach (var projectId in this.ProjectIds)
-            {
-                var project = this.GetProject(projectId);
-                latestVersion = project.Version.GetNewerVersion(latestVersion);
-            }
+        internal SolutionServices Services => _state.Services;
 
-            return latestVersion;
-        }
+        internal BranchId BranchId => _state.BranchId;
 
-        internal int WorkspaceVersion
-        {
-            get { return _workspaceVersion; }
-        }
-
-        internal SolutionServices Services
-        {
-            get { return _solutionServices; }
-        }
-
-        /// <summary>
-        /// branch id of this solution
-        /// 
-        /// currently, it only supports one level of branching. there is a primary branch of a workspace and all other
-        /// branches that are branched from the primary branch.
-        /// 
-        /// one still can create multiple forked solutions from an already branched solution, but versions among those
-        /// can't be reliably used and compared. 
-        /// 
-        /// version only has a meaning between primary solution and branched one or between solutions from same branch.
-        /// </summary>
-        internal BranchId BranchId
-        {
-            get { return _branchId; }
-        }
+        internal ProjectState GetProjectState(ProjectId projectId) => _state.GetProjectState(projectId);
 
         /// <summary>
         /// The Workspace this solution is associated with.
         /// </summary>
-        public Workspace Workspace
-        {
-            get { return _solutionServices.Workspace; }
-        }
+        public Workspace Workspace => _state.Workspace;
 
         /// <summary>
         /// The Id of the solution. Multiple solution instances may share the same Id.
         /// </summary>
-        public SolutionId Id
-        {
-            get { return _id; }
-        }
+        public SolutionId Id => _state.Id;
 
         /// <summary>
         /// The path to the solution file or null if there is no solution file.
         /// </summary>
-        public string FilePath
-        {
-            get { return _filePath; }
-        }
+        public string FilePath => _state.FilePath;
 
         /// <summary>
         /// The solution version. This equates to the solution file's version.
         /// </summary>
-        public VersionStamp Version
-        {
-            get { return _version; }
-        }
+        public VersionStamp Version => _state.Version;
 
         /// <summary>
         /// A list of all the ids for all the projects contained by the solution.
         /// </summary>
-        public IReadOnlyList<ProjectId> ProjectIds
-        {
-            get { return _projectIds; }
-        }
+        public IReadOnlyList<ProjectId> ProjectIds => _state.ProjectIds;
 
         /// <summary>
         /// A list of all the projects contained by the solution.
         /// </summary>
-        public IEnumerable<Project> Projects
-        {
-            get { return _projectIds.Select(id => GetProject(id)); }
-        }
-
-        // [Conditional("DEBUG")]
-        private void CheckInvariants()
-        {
-            Contract.ThrowIfTrue(_projectIds.Count != _projectIdToProjectStateMap.Count);
-
-            // An id shouldn't point at a tracker for a different project.
-            Contract.ThrowIfTrue(_projectIdToTrackerMap.Any(kvp => kvp.Key != kvp.Value.ProjectState.Id));
-        }
-
-        private Solution Branch(
-            IEnumerable<ProjectId> projectIds = null,
-            ImmutableDictionary<ProjectId, ProjectState> idToProjectStateMap = null,
-            ImmutableDictionary<ProjectId, CompilationTracker> projectIdToTrackerMap = null,
-            ImmutableDictionary<string, ImmutableArray<DocumentId>> linkedFilesMap = null,
-            ProjectDependencyGraph dependencyGraph = null,
-            VersionStamp? version = default(VersionStamp?),
-            Lazy<VersionStamp> lazyLatestProjectVersion = null)
-        {
-            var branchId = GetBranchId();
-
-            projectIds = projectIds ?? _projectIds;
-            idToProjectStateMap = idToProjectStateMap ?? _projectIdToProjectStateMap;
-            projectIdToTrackerMap = projectIdToTrackerMap ?? _projectIdToTrackerMap;
-            linkedFilesMap = linkedFilesMap ?? _linkedFilesMap;
-            dependencyGraph = dependencyGraph ?? _dependencyGraph;
-            version = version.HasValue ? version.Value : _version;
-            lazyLatestProjectVersion = lazyLatestProjectVersion ?? _lazyLatestProjectVersion;
-
-            if (branchId == _branchId &&
-                projectIds == _projectIds &&
-                idToProjectStateMap == _projectIdToProjectStateMap &&
-                projectIdToTrackerMap == _projectIdToTrackerMap &&
-                linkedFilesMap == _linkedFilesMap &&
-                dependencyGraph == _dependencyGraph &&
-                version == _version &&
-                lazyLatestProjectVersion == _lazyLatestProjectVersion)
-            {
-                return this;
-            }
-
-            return new Solution(
-                branchId,
-                _workspaceVersion,
-                _solutionServices,
-                _id,
-                _filePath,
-                projectIds,
-                idToProjectStateMap,
-                projectIdToTrackerMap,
-                linkedFilesMap,
-                dependencyGraph,
-                version.Value,
-                lazyLatestProjectVersion);
-        }
-
-        private Solution CreatePrimarySolution(
-            BranchId branchId,
-            int workspaceVersion,
-            SolutionServices services)
-        {
-            if (branchId == _branchId &&
-                workspaceVersion == _workspaceVersion &&
-                services == _solutionServices)
-            {
-                return this;
-            }
-
-            return new Solution(
-                branchId,
-                workspaceVersion,
-                services,
-                _id,
-                _filePath,
-                _projectIds,
-                _projectIdToProjectStateMap,
-                _projectIdToTrackerMap,
-                _linkedFilesMap,
-                _dependencyGraph,
-                _version,
-                _lazyLatestProjectVersion);
-        }
-
-        private BranchId GetBranchId()
-        {
-            // currently we only support one level branching. 
-            // my reasonings are
-            // 1. it seems there is no-one who needs sub branches.
-            // 2. this lets us to branch without explicit branch API
-            return _branchId == Workspace.PrimaryBranchId ? BranchId.GetNextId() : _branchId;
-        }
+        public IEnumerable<Project> Projects => ProjectIds.Select(id => GetProject(id));
 
         /// <summary>
         /// The version of the most recently modified project.
         /// </summary>
-        public VersionStamp GetLatestProjectVersion()
-        {
-            return _lazyLatestProjectVersion.Value;
-        }
+        public VersionStamp GetLatestProjectVersion() => _state.GetLatestProjectVersion();
 
         /// <summary>
         /// True if the solution contains a project with the specified project ID.
         /// </summary>
-        public bool ContainsProject(ProjectId projectId)
-        {
-            return projectId != null && _projectIdToProjectStateMap.ContainsKey(projectId);
-        }
+        public bool ContainsProject(ProjectId projectId) => _state.ContainsProject(projectId);
 
         /// <summary>
         /// Gets the project in this solution with the specified project ID. 
@@ -321,7 +110,7 @@ namespace Microsoft.CodeAnalysis
         private static readonly Func<ProjectId, Solution, Project> s_createProjectFunction = CreateProject;
         private static Project CreateProject(ProjectId projectId, Solution solution)
         {
-            return new Project(solution, solution.GetProjectState(projectId));
+            return new Project(solution, solution.State.GetProjectState(projectId));
         }
 
         /// <summary>
@@ -329,59 +118,25 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Project GetProject(IAssemblySymbol assemblySymbol, CancellationToken cancellationToken = default(CancellationToken))
         {
-            if (assemblySymbol == null)
-            {
-                return null;
-            }
+            var projectState = _state.GetProjectState(assemblySymbol, cancellationToken);
 
-            // TODO: Remove this loop when we add source assembly symbols to s_assemblyOrModuleSymbolToProjectMap
-            foreach (var state in _projectIdToProjectStateMap.Values)
-            {
-                Compilation compilation;
-                if (this.TryGetCompilation(state.Id, out compilation))
-                {
-                    // if the symbol is the compilation's assembly symbol, we are done
-                    if (compilation.Assembly == assemblySymbol)
-                    {
-                        return this.GetProject(state.Id);
-                    }
-                }
-            }
-
-            ProjectId id;
-            s_assemblyOrModuleSymbolToProjectMap.TryGetValue(assemblySymbol, out id);
-            return id == null ? null : this.GetProject(id);
+            return projectState == null ? null : GetProject(projectState.Id);
         }
 
         /// <summary>
         /// True if the solution contains the document in one of its projects
         /// </summary>
-        public bool ContainsDocument(DocumentId documentId)
-        {
-            return
-                documentId != null &&
-                this.ContainsProject(documentId.ProjectId) &&
-                this.GetProjectState(documentId.ProjectId).ContainsDocument(documentId);
-        }
+        public bool ContainsDocument(DocumentId documentId) => _state.ContainsDocument(documentId);
 
         /// <summary>
         /// True if the solution contains the additional document in one of its projects
         /// </summary>
-        public bool ContainsAdditionalDocument(DocumentId documentId)
-        {
-            return
-                documentId != null &&
-                this.ContainsProject(documentId.ProjectId) &&
-                this.GetProjectState(documentId.ProjectId).ContainsAdditionalDocument(documentId);
-        }
+        public bool ContainsAdditionalDocument(DocumentId documentId) => _state.ContainsAdditionalDocument(documentId);
 
         /// <summary>
         /// Gets the documentId in this solution with the specified syntax tree.
         /// </summary>
-        public DocumentId GetDocumentId(SyntaxTree syntaxTree)
-        {
-            return this.GetDocumentId(syntaxTree, null);
-        }
+        public DocumentId GetDocumentId(SyntaxTree syntaxTree) => GetDocumentId(syntaxTree, projectId: null);
 
         /// <summary>
         /// Gets the documentId in this solution with the specified syntax tree.
@@ -431,40 +186,12 @@ namespace Microsoft.CodeAnalysis
             return null;
         }
 
-        private DocumentState GetDocumentState(DocumentId documentId)
-        {
-            if (documentId != null)
-            {
-                var projectState = this.GetProjectState(documentId.ProjectId);
-                if (projectState != null)
-                {
-                    return projectState.GetDocumentState(documentId);
-                }
-            }
-
-            return null;
-        }
-
-        private TextDocumentState GetAdditionalDocumentState(DocumentId documentId)
-        {
-            if (documentId != null)
-            {
-                var projectState = this.GetProjectState(documentId.ProjectId);
-                if (projectState != null)
-                {
-                    return projectState.GetAdditionalDocumentState(documentId);
-                }
-            }
-
-            return null;
-        }
-
         /// <summary>
         /// Gets the document in this solution with the specified syntax tree.
         /// </summary>
         public Document GetDocument(SyntaxTree syntaxTree)
         {
-            return this.GetDocument(syntaxTree, null);
+            return this.GetDocument(syntaxTree, projectId: null);
         }
 
         internal Document GetDocument(SyntaxTree syntaxTree, ProjectId projectId)
@@ -492,63 +219,6 @@ namespace Microsoft.CodeAnalysis
             return null;
         }
 
-        internal Task<VersionStamp> GetDependentVersionAsync(ProjectId projectId, CancellationToken cancellationToken)
-        {
-            return this.GetCompilationTracker(projectId).GetDependentVersionAsync(this, cancellationToken);
-        }
-
-        internal Task<VersionStamp> GetDependentSemanticVersionAsync(ProjectId projectId, CancellationToken cancellationToken)
-        {
-            return this.GetCompilationTracker(projectId).GetDependentSemanticVersionAsync(this, cancellationToken);
-        }
-
-        internal ProjectState GetProjectState(ProjectId projectId)
-        {
-            ProjectState state;
-            _projectIdToProjectStateMap.TryGetValue(projectId, out state);
-            return state;
-        }
-
-        private bool TryGetCompilationTracker(ProjectId projectId, out CompilationTracker tracker)
-        {
-            return _projectIdToTrackerMap.TryGetValue(projectId, out tracker);
-        }
-
-        private static readonly Func<ProjectId, Solution, CompilationTracker> s_createCompilationTrackerFunction = CreateCompilationTracker;
-        private static CompilationTracker CreateCompilationTracker(ProjectId projectId, Solution solution)
-        {
-            return new CompilationTracker(solution.GetProjectState(projectId));
-        }
-
-        private CompilationTracker GetCompilationTracker(ProjectId projectId)
-        {
-            CompilationTracker tracker;
-            if (!_projectIdToTrackerMap.TryGetValue(projectId, out tracker))
-            {
-                tracker = ImmutableInterlocked.GetOrAdd(ref _projectIdToTrackerMap, projectId, s_createCompilationTrackerFunction, this);
-            }
-
-            return tracker;
-        }
-
-        private Solution AddProject(ProjectId projectId, ProjectState projectState)
-        {
-            var newProjectIds = _projectIds.ToImmutableArray().Add(projectId);
-            var newStateMap = _projectIdToProjectStateMap.Add(projectId, projectState);
-            var newDependencyGraph = CreateDependencyGraph(newProjectIds, newStateMap);
-            var newTrackerMap = CreateCompilationTrackerMap(projectId, newDependencyGraph);
-            var newLinkedFilesMap = CreateLinkedFilesMapWithAddedProject(newStateMap[projectId]);
-
-            return this.Branch(
-                projectIds: newProjectIds,
-                idToProjectStateMap: newStateMap,
-                projectIdToTrackerMap: newTrackerMap,
-                linkedFilesMap: newLinkedFilesMap,
-                dependencyGraph: newDependencyGraph,
-                version: this.Version.GetNewerVersion(),  // changed project list so, increment version.
-                lazyLatestProjectVersion: new Lazy<VersionStamp>(() => projectState.Version)); // this is the newest!
-        }
-
         /// <summary>
         /// Creates a new solution instance that includes a project with the specified language and names.
         /// Returns the new project.
@@ -572,63 +242,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution AddProject(ProjectInfo projectInfo)
         {
-            if (projectInfo == null)
+            var newState = _state.AddProject(projectInfo);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectInfo));
+                return this;
             }
 
-            var projectId = projectInfo.Id;
-
-            var language = projectInfo.Language;
-            if (language == null)
-            {
-                throw new ArgumentNullException(nameof(language));
-            }
-
-            var displayName = projectInfo.Name;
-            if (displayName == null)
-            {
-                throw new ArgumentNullException(nameof(displayName));
-            }
-
-            CheckNotContainsProject(projectId);
-
-            var languageServices = this.Workspace.Services.GetLanguageServices(language);
-            if (languageServices == null)
-            {
-                throw new ArgumentException(string.Format(WorkspacesResources.The_language_0_is_not_supported, language));
-            }
-
-            var newProject = new ProjectState(projectInfo, languageServices, _solutionServices);
-
-            return this.AddProject(newProject.Id, newProject);
-        }
-
-        private ImmutableDictionary<string, ImmutableArray<DocumentId>> CreateLinkedFilesMapWithAddedProject(ProjectState projectState)
-        {
-            return CreateLinkedFilesMapWithAddedDocuments(projectState, projectState.DocumentIds);
-        }
-
-        private ImmutableDictionary<string, ImmutableArray<DocumentId>> CreateLinkedFilesMapWithAddedDocuments(ProjectState projectState, IEnumerable<DocumentId> documentIds)
-        {
-            var builder = _linkedFilesMap.ToBuilder();
-
-            foreach (var documentId in documentIds)
-            {
-                var filePath = projectState.GetDocumentState(documentId).FilePath;
-
-                if (string.IsNullOrEmpty(filePath))
-                {
-                    continue;
-                }
-
-                ImmutableArray<DocumentId> documentIdsWithPath;
-                builder[filePath] = builder.TryGetValue(filePath, out documentIdsWithPath)
-                    ? documentIdsWithPath.Add(documentId)
-                    : ImmutableArray.Create(documentId);
-            }
-
-            return builder.ToImmutable();
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -636,65 +256,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution RemoveProject(ProjectId projectId)
         {
-            if (projectId == null)
+            var newState = _state.RemoveProject(projectId);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            CheckContainsProject(projectId);
-
-            var newProjectIds = _projectIds.ToImmutableArray().Remove(projectId);
-            var newStateMap = _projectIdToProjectStateMap.Remove(projectId);
-            var newDependencyGraph = CreateDependencyGraph(newProjectIds, newStateMap);
-            var newTrackerMap = CreateCompilationTrackerMap(projectId, newDependencyGraph);
-            var newLinkedFilesMap = CreateLinkedFilesMapWithRemovedProject(_projectIdToProjectStateMap[projectId]);
-
-            return this.Branch(
-                projectIds: newProjectIds,
-                idToProjectStateMap: newStateMap,
-                projectIdToTrackerMap: newTrackerMap.Remove(projectId),
-                linkedFilesMap: newLinkedFilesMap,
-                dependencyGraph: newDependencyGraph,
-                version: this.Version.GetNewerVersion()); // changed project list, so increment version
-        }
-
-        private ImmutableDictionary<string, ImmutableArray<DocumentId>> CreateLinkedFilesMapWithRemovedProject(ProjectState projectState)
-        {
-            return CreateLinkedFilesMapWithRemovedDocuments(projectState, projectState.DocumentIds);
-        }
-
-        private ImmutableDictionary<string, ImmutableArray<DocumentId>> CreateLinkedFilesMapWithRemovedDocuments(
-            ProjectState projectState,
-            IEnumerable<DocumentId> documentIds)
-        {
-            var builder = _linkedFilesMap.ToBuilder();
-
-            foreach (var documentId in documentIds)
-            {
-                var filePath = projectState.GetDocumentState(documentId).FilePath;
-
-                if (string.IsNullOrEmpty(filePath))
-                {
-                    continue;
-                }
-
-                ImmutableArray<DocumentId> documentIdsWithPath;
-                if (!builder.TryGetValue(filePath, out documentIdsWithPath) || !documentIdsWithPath.Contains(documentId))
-                {
-                    throw new ArgumentException("The given documentId was not found in the linkedFilesMap.");
-                }
-
-                if (documentIdsWithPath.Length == 1)
-                {
-                    builder.Remove(filePath);
-                }
-                else
-                {
-                    builder[filePath] = documentIdsWithPath.Remove(documentId);
-                }
-            }
-
-            return builder.ToImmutable();
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -703,27 +271,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithProjectAssemblyName(ProjectId projectId, string assemblyName)
         {
-            if (projectId == null)
-            {
-                throw new ArgumentNullException(nameof(projectId));
-            }
-
-            if (assemblyName == null)
-            {
-                throw new ArgumentNullException(nameof(assemblyName));
-            }
-
-            CheckContainsProject(projectId);
-
-            var oldProject = this.GetProjectState(projectId);
-            var newProject = oldProject.UpdateAssemblyName(assemblyName);
-
-            if (oldProject == newProject)
+            var newState = _state.WithProjectAssemblyName(projectId, assemblyName);
+            if (newState == _state)
             {
                 return this;
             }
 
-            return this.ForkProject(newProject, CompilationTranslationAction.ProjectAssemblyName(assemblyName));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -731,14 +285,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithProjectOutputFilePath(ProjectId projectId, string outputFilePath)
         {
-            if (projectId == null)
+            var newState = _state.WithProjectOutputFilePath(projectId, outputFilePath);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            CheckContainsProject(projectId);
-
-            return this.ForkProject(this.GetProjectState(projectId).UpdateOutputPath(outputFilePath));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -746,14 +299,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithProjectName(ProjectId projectId, string name)
         {
-            if (projectId == null)
+            var newState = _state.WithProjectName(projectId, name);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            CheckContainsProject(projectId);
-
-            return this.ForkProject(this.GetProjectState(projectId).UpdateName(name));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -761,14 +313,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithProjectFilePath(ProjectId projectId, string filePath)
         {
-            if (projectId == null)
+            var newState = _state.WithProjectFilePath(projectId, filePath);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            CheckContainsProject(projectId);
-
-            return this.ForkProject(this.GetProjectState(projectId).UpdateFilePath(filePath));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -777,27 +328,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithProjectCompilationOptions(ProjectId projectId, CompilationOptions options)
         {
-            if (projectId == null)
-            {
-                throw new ArgumentNullException(nameof(projectId));
-            }
-
-            if (options == null)
-            {
-                throw new ArgumentNullException(nameof(options));
-            }
-
-            CheckContainsProject(projectId);
-
-            var oldProject = this.GetProjectState(projectId);
-            var newProject = oldProject.UpdateCompilationOptions(options);
-
-            if (oldProject == newProject)
+            var newState = _state.WithProjectCompilationOptions(projectId, options);
+            if (newState == _state)
             {
                 return this;
             }
 
-            return this.ForkProject(newProject, CompilationTranslationAction.ProjectCompilationOptions(options));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -806,31 +343,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithProjectParseOptions(ProjectId projectId, ParseOptions options)
         {
-            if (projectId == null)
-            {
-                throw new ArgumentNullException(nameof(projectId));
-            }
-
-            Contract.Requires(this.ContainsProject(projectId));
-
-            var oldProject = this.GetProjectState(projectId);
-            var newProject = oldProject.UpdateParseOptions(options);
-
-            if (oldProject == newProject)
+            var newState = _state.WithProjectParseOptions(projectId, options);
+            if (newState == _state)
             {
                 return this;
             }
 
-            if (this.Workspace.PartialSemanticsEnabled)
-            {
-                // don't fork tracker with queued action since access via partial semantics can become inconsistent (throw).
-                // Since changing options is rare event, it is okay to start compilation building from scratch.
-                return this.ForkProject(newProject, forkTracker: false);
-            }
-            else
-            {
-                return this.ForkProject(newProject, CompilationTranslationAction.ProjectParseOptions(newProject));
-            }
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -840,36 +359,13 @@ namespace Microsoft.CodeAnalysis
         // TODO: make it public
         internal Solution WithHasAllInformation(ProjectId projectId, bool hasAllInformation)
         {
-            if (projectId == null)
-            {
-                throw new ArgumentNullException(nameof(projectId));
-            }
-
-            Contract.Requires(this.ContainsProject(projectId));
-
-            var oldProject = this.GetProjectState(projectId);
-            var newProject = oldProject.UpdateHasAllInformation(hasAllInformation);
-
-            if (oldProject == newProject)
+            var newState = _state.WithHasAllInformation(projectId, hasAllInformation);
+            if (newState == _state)
             {
                 return this;
             }
 
-            // fork without any change on compilation.
-            return this.ForkProject(newProject);
-        }
-
-        private static async Task<Compilation> ReplaceSyntaxTreesWithTreesFromNewProjectStateAsync(Compilation compilation, ProjectState projectState, CancellationToken cancellationToken)
-        {
-            var syntaxTrees = new List<SyntaxTree>(capacity: projectState.DocumentIds.Count);
-
-            foreach (var documentState in projectState.OrderedDocumentStates)
-            {
-                cancellationToken.ThrowIfCancellationRequested();
-                syntaxTrees.Add(await documentState.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false));
-            }
-
-            return compilation.RemoveAllSyntaxTrees().AddSyntaxTrees(syntaxTrees);
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -878,27 +374,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution AddProjectReference(ProjectId projectId, ProjectReference projectReference)
         {
-            if (projectId == null)
+            var newState = _state.AddProjectReference(projectId, projectReference);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            if (projectReference == null)
-            {
-                throw new ArgumentNullException(nameof(projectReference));
-            }
-
-            CheckContainsProject(projectId);
-            CheckContainsProject(projectReference.ProjectId);
-            CheckNotContainsProjectReference(projectId, projectReference);
-            CheckNotContainsTransitiveReference(projectReference.ProjectId, projectId);
-
-            CheckNotSecondSubmissionReference(projectId, projectReference.ProjectId);
-
-            var oldProject = this.GetProjectState(projectId);
-            var newProject = oldProject.AddProjectReference(projectReference);
-
-            return this.ForkProject(newProject, withProjectReferenceChange: true);
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -907,30 +389,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution AddProjectReferences(ProjectId projectId, IEnumerable<ProjectReference> projectReferences)
         {
-            if (projectId == null)
+            var newState = _state.AddProjectReferences(projectId, projectReferences);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            if (projectReferences == null)
-            {
-                throw new ArgumentNullException(nameof(projectReferences));
-            }
-
-            CheckContainsProject(projectId);
-
-            foreach (var referencedProject in projectReferences)
-            {
-                CheckContainsProject(referencedProject.ProjectId);
-                CheckNotContainsProjectReference(projectId, referencedProject);
-                CheckNotContainsTransitiveReference(referencedProject.ProjectId, projectId);
-                CheckNotSecondSubmissionReference(projectId, referencedProject.ProjectId);
-            }
-
-            var oldProject = this.GetProjectState(projectId);
-            var newProject = oldProject.AddProjectReferences(projectReferences);
-
-            return this.ForkProject(newProject, withProjectReferenceChange: true);
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -939,23 +404,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution RemoveProjectReference(ProjectId projectId, ProjectReference projectReference)
         {
-            if (projectId == null)
+            var newState = _state.RemoveProjectReference(projectId, projectReference);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            if (projectReference == null)
-            {
-                throw new ArgumentNullException(nameof(projectReference));
-            }
-
-            CheckContainsProject(projectId);
-            CheckContainsProject(projectReference.ProjectId);
-
-            var oldProject = this.GetProjectState(projectId);
-            var newProject = oldProject.RemoveProjectReference(projectReference);
-
-            return this.ForkProject(newProject, withProjectReferenceChange: true);
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -964,22 +419,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithProjectReferences(ProjectId projectId, IEnumerable<ProjectReference> projectReferences)
         {
-            if (projectId == null)
+            var newState = _state.WithProjectReferences(projectId, projectReferences);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            if (projectReferences == null)
-            {
-                throw new ArgumentNullException(nameof(projectReferences));
-            }
-
-            CheckContainsProject(projectId);
-
-            var oldProject = this.GetProjectState(projectId);
-            var newProject = oldProject.WithProjectReferences(projectReferences);
-
-            return this.ForkProject(newProject, withProjectReferenceChange: true);
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -988,20 +434,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution AddMetadataReference(ProjectId projectId, MetadataReference metadataReference)
         {
-            if (projectId == null)
+            var newState = _state.AddMetadataReference(projectId, metadataReference);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            if (metadataReference == null)
-            {
-                throw new ArgumentNullException(nameof(metadataReference));
-            }
-
-            CheckContainsProject(projectId);
-
-            return this.ForkProject(
-                this.GetProjectState(projectId).AddMetadataReference(metadataReference));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1010,18 +449,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution AddMetadataReferences(ProjectId projectId, IEnumerable<MetadataReference> metadataReferences)
         {
-            if (projectId == null)
+            var newState = _state.AddMetadataReferences(projectId, metadataReferences);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            if (metadataReferences == null)
-            {
-                throw new ArgumentNullException(nameof(metadataReferences));
-            }
-
-            CheckContainsProject(projectId);
-            return this.ForkProject(this.GetProjectState(projectId).AddMetadataReferences(metadataReferences));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1030,20 +464,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution RemoveMetadataReference(ProjectId projectId, MetadataReference metadataReference)
         {
-            if (projectId == null)
+            var newState = _state.RemoveMetadataReference(projectId, metadataReference);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            if (metadataReference == null)
-            {
-                throw new ArgumentNullException(nameof(metadataReference));
-            }
-
-            CheckContainsProject(projectId);
-
-            return this.ForkProject(
-                this.GetProjectState(projectId).RemoveMetadataReference(metadataReference));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1052,18 +479,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithProjectMetadataReferences(ProjectId projectId, IEnumerable<MetadataReference> metadataReferences)
         {
-            if (projectId == null)
+            var newState = _state.WithProjectMetadataReferences(projectId, metadataReferences);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            if (metadataReferences == null)
-            {
-                throw new ArgumentNullException(nameof(metadataReferences));
-            }
-
-            CheckContainsProject(projectId);
-            return this.ForkProject(this.GetProjectState(projectId).WithMetadataReferences(metadataReferences));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1072,18 +494,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution AddAnalyzerReference(ProjectId projectId, AnalyzerReference analyzerReference)
         {
-            if (projectId == null)
+            var newState = _state.AddAnalyzerReference(projectId, analyzerReference);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            if (analyzerReference == null)
-            {
-                throw new ArgumentNullException(nameof(analyzerReference));
-            }
-
-            CheckContainsProject(projectId);
-            return this.ForkProject(this.GetProjectState(projectId).AddAnalyzerReference(analyzerReference));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1092,18 +509,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution AddAnalyzerReferences(ProjectId projectId, IEnumerable<AnalyzerReference> analyzerReferences)
         {
-            if (projectId == null)
+            var newState = _state.AddAnalyzerReferences(projectId, analyzerReferences);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            if (analyzerReferences == null)
-            {
-                throw new ArgumentNullException(nameof(analyzerReferences));
-            }
-
-            CheckContainsProject(projectId);
-            return this.ForkProject(this.GetProjectState(projectId).AddAnalyzerReferences(analyzerReferences));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1112,18 +524,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution RemoveAnalyzerReference(ProjectId projectId, AnalyzerReference analyzerReference)
         {
-            if (projectId == null)
+            var newState = _state.RemoveAnalyzerReference(projectId, analyzerReference);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            if (analyzerReference == null)
-            {
-                throw new ArgumentNullException(nameof(analyzerReference));
-            }
-
-            CheckContainsProject(projectId);
-            return this.ForkProject(this.GetProjectState(projectId).RemoveAnalyzerReference(analyzerReference));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1132,36 +539,18 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithProjectAnalyzerReferences(ProjectId projectId, IEnumerable<AnalyzerReference> analyzerReferences)
         {
-            if (projectId == null)
+            var newState = _state.WithProjectAnalyzerReferences(projectId, analyzerReferences);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(projectId));
+                return this;
             }
 
-            if (analyzerReferences == null)
-            {
-                throw new ArgumentNullException(nameof(analyzerReferences));
-            }
-
-            CheckContainsProject(projectId);
-            return this.ForkProject(this.GetProjectState(projectId).WithAnalyzerReferences(analyzerReferences));
+            return new Solution(newState);
         }
 
-        private Solution AddDocument(DocumentState state)
+        private static SourceCodeKind GetSourceCodeKind(ProjectState project)
         {
-            if (state == null)
-            {
-                throw new ArgumentNullException(nameof(state));
-            }
-
-            CheckContainsProject(state.Id.ProjectId);
-
-            var oldProject = this.GetProjectState(state.Id.ProjectId);
-            var newProject = oldProject.AddDocument(state);
-
-            return this.ForkProject(
-                newProject,
-                CompilationTranslationAction.AddDocument(state),
-                newLinkedFilesMap: CreateLinkedFilesMapWithAddedDocuments(newProject, SpecializedCollections.SingletonEnumerable(state.Id)));
+            return project.ParseOptions != null ? project.ParseOptions.Kind : SourceCodeKind.Regular;
         }
 
         /// <summary>
@@ -1184,9 +573,6 @@ namespace Microsoft.CodeAnalysis
                 throw new ArgumentNullException(nameof(documentId));
             }
 
-            CheckContainsProject(documentId.ProjectId);
-            CheckNotContainsDocument(documentId);
-
             if (name == null)
             {
                 throw new ArgumentNullException(nameof(name));
@@ -1197,7 +583,7 @@ namespace Microsoft.CodeAnalysis
                 throw new ArgumentNullException(nameof(text));
             }
 
-            var project = this.GetProjectState(documentId.ProjectId);
+            var project = _state.GetProjectState(documentId.ProjectId);
 
             var version = VersionStamp.Create();
             var loader = TextLoader.From(TextAndVersion.Create(text, version, name));
@@ -1211,18 +597,7 @@ namespace Microsoft.CodeAnalysis
                 filePath: filePath,
                 isGenerated: isGenerated);
 
-            var doc = DocumentState.Create(
-                info,
-                project.ParseOptions,
-                project.LanguageServices,
-                _solutionServices);
-
-            return this.AddDocument(doc);
-        }
-
-        private static SourceCodeKind GetSourceCodeKind(ProjectState project)
-        {
-            return project.ParseOptions != null ? project.ParseOptions.Kind : SourceCodeKind.Regular;
+            return this.AddDocument(info);
         }
 
         /// <summary>
@@ -1231,7 +606,7 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution AddDocument(DocumentId documentId, string name, SyntaxNode syntaxRoot, IEnumerable<string> folders = null, string filePath = null, bool isGenerated = false, PreservationMode preservationMode = PreservationMode.PreserveValue)
         {
-            return AddDocument(documentId, name, SourceText.From(string.Empty), folders, filePath, isGenerated).WithDocumentSyntaxRoot(documentId, syntaxRoot, preservationMode);
+            return this.AddDocument(documentId, name, SourceText.From(string.Empty), folders, filePath, isGenerated).WithDocumentSyntaxRoot(documentId, syntaxRoot, preservationMode);
         }
 
         /// <summary>
@@ -1245,9 +620,6 @@ namespace Microsoft.CodeAnalysis
                 throw new ArgumentNullException(nameof(documentId));
             }
 
-            CheckContainsProject(documentId.ProjectId);
-            CheckNotContainsDocument(documentId);
-
             if (name == null)
             {
                 throw new ArgumentNullException(nameof(name));
@@ -1258,7 +630,7 @@ namespace Microsoft.CodeAnalysis
                 throw new ArgumentNullException(nameof(loader));
             }
 
-            var project = this.GetProjectState(documentId.ProjectId);
+            var project = _state.GetProjectState(documentId.ProjectId);
 
             var info = DocumentInfo.Create(
                 documentId,
@@ -1276,36 +648,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution AddDocument(DocumentInfo documentInfo)
         {
-            if (documentInfo == null)
+            var newState = _state.AddDocument(documentInfo);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(documentInfo));
+                return this;
             }
 
-            CheckContainsProject(documentInfo.Id.ProjectId);
-            CheckNotContainsDocument(documentInfo.Id);
-
-            var project = this.GetProjectState(documentInfo.Id.ProjectId);
-
-            var doc = DocumentState.Create(
-                documentInfo,
-                project.ParseOptions,
-                project.LanguageServices,
-                _solutionServices).UpdateSourceCodeKind(documentInfo.SourceCodeKind);
-
-            return this.AddDocument(doc);
-        }
-
-        private Solution AddDocument(Document document)
-        {
-            if (document == null)
-            {
-                throw new ArgumentNullException(nameof(document));
-            }
-
-            CheckNotContainsDocument(document.Id);
-            CheckContainsProject(document.Id.ProjectId);
-
-            return this.AddDocument((DocumentState)document.State);
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1328,9 +677,6 @@ namespace Microsoft.CodeAnalysis
                 throw new ArgumentNullException(nameof(documentId));
             }
 
-            CheckContainsProject(documentId.ProjectId);
-            CheckNotContainsAdditionalDocument(documentId);
-
             if (name == null)
             {
                 throw new ArgumentNullException(nameof(name));
@@ -1341,7 +687,7 @@ namespace Microsoft.CodeAnalysis
                 throw new ArgumentNullException(nameof(text));
             }
 
-            var project = this.GetProjectState(documentId.ProjectId);
+            var project = _state.GetProjectState(documentId.ProjectId);
 
             var version = VersionStamp.Create();
             var loader = TextLoader.From(TextAndVersion.Create(text, version, name));
@@ -1354,45 +700,18 @@ namespace Microsoft.CodeAnalysis
                 loader: loader,
                 filePath: filePath);
 
-            var doc = TextDocumentState.Create(
-                info,
-                _solutionServices);
-
-            return this.AddAdditionalDocument(doc);
+            return this.AddAdditionalDocument(info);
         }
 
         public Solution AddAdditionalDocument(DocumentInfo documentInfo)
         {
-            if (documentInfo == null)
+            var newState = _state.AddAdditionalDocument(documentInfo);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(documentInfo));
+                return this;
             }
 
-            CheckContainsProject(documentInfo.Id.ProjectId);
-            CheckNotContainsAdditionalDocument(documentInfo.Id);
-
-            var project = this.GetProjectState(documentInfo.Id.ProjectId);
-
-            var doc = TextDocumentState.Create(
-                documentInfo,
-                _solutionServices);
-
-            return this.AddAdditionalDocument(doc);
-        }
-
-        private Solution AddAdditionalDocument(TextDocumentState state)
-        {
-            if (state == null)
-            {
-                throw new ArgumentNullException(nameof(state));
-            }
-
-            CheckContainsProject(state.Id.ProjectId);
-
-            var oldProject = this.GetProjectState(state.Id.ProjectId);
-            var newProject = oldProject.AddAdditionalDocument(state);
-
-            return this.ForkProject(newProject);
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1400,16 +719,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution RemoveDocument(DocumentId documentId)
         {
-            CheckContainsDocument(documentId);
+            var newState = _state.RemoveDocument(documentId);
+            if (newState == _state)
+            {
+                return this;
+            }
 
-            var oldProject = this.GetProjectState(documentId.ProjectId);
-            var oldDocument = oldProject.GetDocumentState(documentId);
-            var newProject = oldProject.RemoveDocument(documentId);
-
-            return this.ForkProject(
-                newProject,
-                CompilationTranslationAction.RemoveDocument(oldDocument),
-                newLinkedFilesMap: CreateLinkedFilesMapWithRemovedDocuments(oldProject, SpecializedCollections.SingletonEnumerable(documentId)));
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1417,12 +733,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution RemoveAdditionalDocument(DocumentId documentId)
         {
-            CheckContainsAdditionalDocument(documentId);
+            var newState = _state.RemoveAdditionalDocument(documentId);
+            if (newState == _state)
+            {
+                return this;
+            }
 
-            var oldProject = this.GetProjectState(documentId.ProjectId);
-            var newProject = oldProject.RemoveAdditionalDocument(documentId);
-
-            return this.ForkProject(newProject);
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1431,22 +748,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithDocumentFolders(DocumentId documentId, IEnumerable<string> folders)
         {
-            if (documentId == null)
+            var newState = _state.WithDocumentFolders(documentId, folders);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(documentId));
+                return this;
             }
 
-            if (folders == null)
-            {
-                throw new ArgumentNullException(nameof(folders));
-            }
-
-            folders = folders != null ? folders.WhereNotNull().ToReadOnlyCollection() : null;
-
-            var oldDocument = this.GetDocumentState(documentId);
-            var newDocument = oldDocument.UpdateFolders(folders.WhereNotNull().ToReadOnlyCollection());
-
-            return this.WithDocumentState(newDocument);
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1455,45 +763,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithDocumentText(DocumentId documentId, SourceText text, PreservationMode mode = PreservationMode.PreserveValue)
         {
-            if (documentId == null)
-            {
-                throw new ArgumentNullException(nameof(documentId));
-            }
-
-            if (text == null)
-            {
-                throw new ArgumentNullException(nameof(text));
-            }
-
-            CheckContainsDocument(documentId);
-
-            var oldDocument = this.GetDocumentState(documentId);
-
-            SourceText oldText;
-            if (oldDocument.TryGetText(out oldText) && text == oldText)
+            var newState = _state.WithDocumentText(documentId, text, mode);
+            if (newState == _state)
             {
                 return this;
             }
 
-            // check to see if this solution has already been branched before with the same doc & text changes.
-            // this helps reduce duplicate parsing when typing, and separate services generating duplicate symbols.
-            if (mode == PreservationMode.PreserveIdentity)
-            {
-                var branch = _firstBranch;
-                if (branch != null && branch.Id == documentId && branch.Text == text)
-                {
-                    return branch.Solution;
-                }
-            }
-
-            var newSolution = this.WithDocumentState(oldDocument.UpdateText(text, mode), textChanged: true);
-
-            if (mode == PreservationMode.PreserveIdentity && _firstBranch == null)
-            {
-                Interlocked.CompareExchange(ref _firstBranch, new SolutionBranch(documentId, text, newSolution), null);
-            }
-
-            return newSolution;
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1502,28 +778,122 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithAdditionalDocumentText(DocumentId documentId, SourceText text, PreservationMode mode = PreservationMode.PreserveValue)
         {
-            if (documentId == null)
-            {
-                throw new ArgumentNullException(nameof(documentId));
-            }
-
-            if (text == null)
-            {
-                throw new ArgumentNullException(nameof(text));
-            }
-
-            CheckContainsAdditionalDocument(documentId);
-
-            var oldDocument = this.GetAdditionalDocumentState(documentId);
-
-            SourceText oldText;
-            if (oldDocument.TryGetText(out oldText) && text == oldText)
+            var newState = _state.WithAdditionalDocumentText(documentId, text, mode);
+            if (newState == _state)
             {
                 return this;
             }
 
-            var newSolution = this.WithTextDocumentState(oldDocument.UpdateText(text, mode), textChanged: true);
-            return newSolution;
+            return new Solution(newState);
+        }
+
+        /// <summary>
+        /// Creates a new solution instance with the document specified updated to have the text
+        /// and version specified.
+        /// </summary>
+        public Solution WithDocumentText(DocumentId documentId, TextAndVersion textAndVersion, PreservationMode mode = PreservationMode.PreserveValue)
+        {
+            var newState = _state.WithDocumentText(documentId, textAndVersion, mode);
+            if (newState == _state)
+            {
+                return this;
+            }
+
+            return new Solution(newState);
+        }
+
+        /// <summary>
+        /// Creates a new solution instance with the additional document specified updated to have the text
+        /// and version specified.
+        /// </summary>
+        public Solution WithAdditionalDocumentText(DocumentId documentId, TextAndVersion textAndVersion, PreservationMode mode = PreservationMode.PreserveValue)
+        {
+            var newState = _state.WithAdditionalDocumentText(documentId, textAndVersion, mode);
+            if (newState == _state)
+            {
+                return this;
+            }
+
+            return new Solution(newState);
+        }
+
+        /// <summary>
+        /// Creates a new solution instance with the document specified updated to have a syntax tree
+        /// rooted by the specified syntax node.
+        /// </summary>
+        public Solution WithDocumentSyntaxRoot(DocumentId documentId, SyntaxNode root, PreservationMode mode = PreservationMode.PreserveValue)
+        {
+            var newState = _state.WithDocumentSyntaxRoot(documentId, root, mode);
+            if (newState == _state)
+            {
+                return this;
+            }
+
+            return new Solution(newState);
+        }
+
+        /// <summary>
+        /// Creates a new solution instance with the document specified updated to have the source
+        /// code kind specified.
+        /// </summary>
+        public Solution WithDocumentSourceCodeKind(DocumentId documentId, SourceCodeKind sourceCodeKind)
+        {
+            var newState = _state.WithDocumentSourceCodeKind(documentId, sourceCodeKind);
+            if (newState == _state)
+            {
+                return this;
+            }
+
+            return new Solution(newState);
+        }
+
+        /// <summary>
+        /// Creates a new solution instance with the document specified updated to have the text
+        /// supplied by the text loader.
+        /// </summary>
+        public Solution WithDocumentTextLoader(DocumentId documentId, TextLoader loader, PreservationMode mode)
+        {
+            return WithDocumentTextLoader(documentId, loader, textOpt: null, mode: mode);
+        }
+
+        internal Solution WithDocumentTextLoader(DocumentId documentId, TextLoader loader, SourceText textOpt, PreservationMode mode)
+        {
+            var newState = _state.WithDocumentTextLoader(documentId, loader, textOpt, mode);
+            if (newState == _state)
+            {
+                return this;
+            }
+
+            return new Solution(newState);
+        }
+
+        /// <summary>
+        /// Creates a new solution instance with the additional document specified updated to have the text
+        /// supplied by the text loader.
+        /// </summary>
+        public Solution WithAdditionalDocumentTextLoader(DocumentId documentId, TextLoader loader, PreservationMode mode)
+        {
+            var newState = _state.WithAdditionalDocumentTextLoader(documentId, loader, mode);
+            if (newState == _state)
+            {
+                return this;
+            }
+
+            return new Solution(newState);
+        }
+
+        /// <summary>
+        /// Creates a branch of the solution that has its compilations frozen in whatever state they are in at the time, assuming a background compiler is
+        /// busy building this compilations.
+        /// 
+        /// A compilation for the project containing the specified document id will be guaranteed to exist with at least the syntax tree for the document.
+        /// 
+        /// This not intended to be the public API, use Document.WithFrozenPartialSemantics() instead.
+        /// </summary>
+        internal async Task<Solution> WithFrozenPartialCompilationIncludingSpecificDocumentAsync(DocumentId documentId, CancellationToken cancellationToken)
+        {
+            var newState = await _state.WithFrozenPartialCompilationIncludingSpecificDocumentAsync(documentId, cancellationToken).ConfigureAwait(false);
+            return new Solution(newState);
         }
 
         internal async Task<Solution> WithMergedLinkedFileChangesAsync(
@@ -1538,280 +908,9 @@ namespace Microsoft.CodeAnalysis
             return (await session.MergeDiffsAsync(mergeConflictHandler, cancellationToken).ConfigureAwait(false)).MergedSolution;
         }
 
-        private SolutionBranch _firstBranch;
-
-        private class SolutionBranch
-        {
-            public readonly DocumentId Id;
-            public readonly SourceText Text;
-            public readonly Solution Solution;
-
-            public SolutionBranch(DocumentId id, SourceText text, Solution solution)
-            {
-                this.Id = id;
-                this.Text = text;
-                this.Solution = solution;
-            }
-        }
-
-        /// <summary>
-        /// Creates a new solution instance with the document specified updated to have the text
-        /// and version specified.
-        /// </summary>
-        public Solution WithDocumentText(DocumentId documentId, TextAndVersion textAndVersion, PreservationMode mode = PreservationMode.PreserveValue)
-        {
-            if (documentId == null)
-            {
-                throw new ArgumentNullException(nameof(documentId));
-            }
-
-            if (textAndVersion == null)
-            {
-                throw new ArgumentNullException(nameof(textAndVersion));
-            }
-
-            CheckContainsDocument(documentId);
-
-            var oldDocument = this.GetDocumentState(documentId);
-
-            return WithDocumentState(oldDocument.UpdateText(textAndVersion, mode), textChanged: true);
-        }
-
-        /// <summary>
-        /// Creates a new solution instance with the additional document specified updated to have the text
-        /// and version specified.
-        /// </summary>
-        public Solution WithAdditionalDocumentText(DocumentId documentId, TextAndVersion textAndVersion, PreservationMode mode = PreservationMode.PreserveValue)
-        {
-            if (documentId == null)
-            {
-                throw new ArgumentNullException(nameof(documentId));
-            }
-
-            if (textAndVersion == null)
-            {
-                throw new ArgumentNullException(nameof(textAndVersion));
-            }
-
-            CheckContainsAdditionalDocument(documentId);
-
-            var oldDocument = this.GetAdditionalDocumentState(documentId);
-
-            return WithTextDocumentState(oldDocument.UpdateText(textAndVersion, mode), textChanged: true);
-        }
-
-        /// <summary>
-        /// Creates a new solution instance with the document specified updated to have a syntax tree
-        /// rooted by the specified syntax node.
-        /// </summary>
-        public Solution WithDocumentSyntaxRoot(DocumentId documentId, SyntaxNode root, PreservationMode mode = PreservationMode.PreserveValue)
-        {
-            if (documentId == null)
-            {
-                throw new ArgumentNullException(nameof(documentId));
-            }
-
-            if (root == null)
-            {
-                throw new ArgumentNullException(nameof(root));
-            }
-
-            CheckContainsDocument(documentId);
-
-            var oldDocument = this.GetDocumentState(documentId);
-
-            SyntaxTree oldTree;
-            SyntaxNode oldRoot;
-            if (oldDocument.TryGetSyntaxTree(out oldTree) &&
-                oldTree.TryGetRoot(out oldRoot) &&
-                oldRoot == root)
-            {
-                return this;
-            }
-
-            return WithDocumentState(oldDocument.UpdateTree(root, mode), textChanged: true);
-        }
-
-        private static async Task<Compilation> UpdateDocumentInCompilationAsync(
-            Compilation compilation,
-            DocumentState oldDocument,
-            DocumentState newDocument,
-            CancellationToken cancellationToken)
-        {
-            return compilation.ReplaceSyntaxTree(
-                await oldDocument.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false),
-                await newDocument.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false));
-        }
-
-        /// <summary>
-        /// Creates a new solution instance with the document specified updated to have the source
-        /// code kind specified.
-        /// </summary>
-        public Solution WithDocumentSourceCodeKind(DocumentId documentId, SourceCodeKind sourceCodeKind)
-        {
-            if (!Enum.IsDefined(typeof(SourceCodeKind), sourceCodeKind))
-            {
-                throw new ArgumentOutOfRangeException(nameof(sourceCodeKind));
-            }
-
-            CheckContainsDocument(documentId);
-
-            var oldDocument = this.GetDocumentState(documentId);
-
-            if (oldDocument.SourceCodeKind == sourceCodeKind)
-            {
-                return this;
-            }
-
-            return WithDocumentState(oldDocument.UpdateSourceCodeKind(sourceCodeKind), textChanged: true);
-        }
-
-        /// <summary>
-        /// Creates a new solution instance with the document specified updated to have the text
-        /// supplied by the text loader.
-        /// </summary>
-        public Solution WithDocumentTextLoader(DocumentId documentId, TextLoader loader, PreservationMode mode)
-        {
-            return WithDocumentTextLoader(documentId, loader, textOpt: null, mode: mode);
-        }
-
-        internal Solution WithDocumentTextLoader(DocumentId documentId, TextLoader loader, SourceText textOpt, PreservationMode mode)
-        {
-            CheckContainsDocument(documentId);
-
-            var oldDocument = this.GetDocumentState(documentId);
-
-            // assumes that text has changed. user could have closed a doc without saving and we are loading text from closed file with
-            // old content. also this should make sure we don't re-use latest doc version with data associated with opened document.
-            return this.WithDocumentState(oldDocument.UpdateText(loader, textOpt, mode), textChanged: true, recalculateDependentVersions: true);
-        }
-
-        /// <summary>
-        /// Creates a new solution instance with the additional document specified updated to have the text
-        /// supplied by the text loader.
-        /// </summary>
-        public Solution WithAdditionalDocumentTextLoader(DocumentId documentId, TextLoader loader, PreservationMode mode)
-        {
-            CheckContainsAdditionalDocument(documentId);
-
-            var oldDocument = this.GetAdditionalDocumentState(documentId);
-
-            // assumes that text has changed. user could have closed a doc without saving and we are loading text from closed file with
-            // old content. also this should make sure we don't re-use latest doc version with data associated with opened document.
-            return this.WithTextDocumentState(oldDocument.UpdateText(loader, mode), textChanged: true, recalculateDependentVersions: true);
-        }
-
-        private Solution WithDocumentState(DocumentState newDocument, bool textChanged = false, bool recalculateDependentVersions = false)
-        {
-            if (newDocument == null)
-            {
-                throw new ArgumentNullException(nameof(newDocument));
-            }
-
-            CheckContainsDocument(newDocument.Id);
-
-            if (newDocument == this.GetDocumentState(newDocument.Id))
-            {
-                // old and new documents are the same instance
-                return this;
-            }
-
-            return this.TouchDocument(newDocument.Id, p => p.UpdateDocument(newDocument, textChanged, recalculateDependentVersions));
-        }
-
-        private Solution TouchDocument(DocumentId documentId, Func<ProjectState, ProjectState> touchProject)
-        {
-            Contract.Requires(this.ContainsDocument(documentId));
-
-            var oldProject = this.GetProjectState(documentId.ProjectId);
-            var newProject = touchProject(oldProject);
-
-            if (oldProject == newProject)
-            {
-                // old and new projects are the same instance
-                return this;
-            }
-
-            var oldDocument = oldProject.GetDocumentState(documentId);
-            var newDocument = newProject.GetDocumentState(documentId);
-
-            return this.ForkProject(newProject, CompilationTranslationAction.TouchDocument(oldDocument, newDocument));
-        }
-
-        private Solution WithTextDocumentState(TextDocumentState newDocument, bool textChanged = false, bool recalculateDependentVersions = false)
-        {
-            if (newDocument == null)
-            {
-                throw new ArgumentNullException(nameof(newDocument));
-            }
-
-            CheckContainsAdditionalDocument(newDocument.Id);
-
-            if (newDocument == this.GetAdditionalDocumentState(newDocument.Id))
-            {
-                // old and new documents are the same instance
-                return this;
-            }
-
-            var oldProject = this.GetProjectState(newDocument.Id.ProjectId);
-            var newProject = oldProject.UpdateAdditionalDocument(newDocument, textChanged, recalculateDependentVersions);
-
-            if (oldProject == newProject)
-            {
-                // old and new projects are the same instance
-                return this;
-            }
-
-            return this.ForkProject(newProject);
-        }
-
-
-        /// <summary>
-        /// Creates a new snapshot with an updated project and an action that will produce a new
-        /// compilation matching the new project out of an old compilation. All dependent projects
-        /// are fixed-up if the change to the new project affects its public metadata, and old
-        /// dependent compilations are forgotten.
-        /// </summary>
-        private Solution ForkProject(
-            ProjectState newProjectState,
-            CompilationTranslationAction translate = null,
-            bool withProjectReferenceChange = false,
-            ImmutableDictionary<string, ImmutableArray<DocumentId>> newLinkedFilesMap = null,
-            bool forkTracker = true)
-        {
-            var projectId = newProjectState.Id;
-
-            var newStateMap = _projectIdToProjectStateMap.SetItem(projectId, newProjectState);
-            var newDependencyGraph = withProjectReferenceChange ? CreateDependencyGraph(_projectIds, newStateMap) : _dependencyGraph;
-            var newTrackerMap = CreateCompilationTrackerMap(projectId, newDependencyGraph);
-
-            // If we have a tracker for this project, then fork it as well (along with the
-            // translation action and store it in the tracker map.
-            CompilationTracker tracker;
-            if (newTrackerMap.TryGetValue(projectId, out tracker))
-            {
-                newTrackerMap = newTrackerMap.Remove(projectId);
-
-                if (forkTracker)
-                {
-                    newTrackerMap = newTrackerMap.Add(projectId, tracker.Fork(newProjectState, translate));
-                }
-            }
-
-            var modifiedDocumentOnly = translate is CompilationTranslationAction.TouchDocumentAction;
-            var newLatestProjectVersion = modifiedDocumentOnly ? _lazyLatestProjectVersion : new Lazy<VersionStamp>(() => newProjectState.Version);
-
-            return this.Branch(
-                idToProjectStateMap: newStateMap,
-                projectIdToTrackerMap: newTrackerMap,
-                dependencyGraph: newDependencyGraph,
-                linkedFilesMap: newLinkedFilesMap ?? _linkedFilesMap,
-                lazyLatestProjectVersion: newLatestProjectVersion);
-        }
-
         internal ImmutableArray<DocumentId> GetRelatedDocumentIds(DocumentId documentId)
         {
-            var projectState = this.GetProjectState(documentId.ProjectId);
+            var projectState = _state.GetProjectState(documentId.ProjectId);
             if (projectState == null)
             {
                 // this document no longer exist
@@ -1836,55 +935,15 @@ namespace Microsoft.CodeAnalysis
             return this.FilterDocumentIdsByLanguage(documentIds, projectState.ProjectInfo.Language).ToImmutableArray();
         }
 
-        /// <summary>
-        /// Gets the set of <see cref="DocumentId"/>s in this <see cref="Solution"/> with a
-        /// <see cref="TextDocument.FilePath"/> that matches the given file path.
-        /// </summary>
-        public ImmutableArray<DocumentId> GetDocumentIdsWithFilePath(string filePath)
+        internal Solution WithNewWorkspace(Workspace workspace, int workspaceVersion)
         {
-            if (string.IsNullOrEmpty(filePath))
+            var newState = _state.WithNewWorkspace(workspace, workspaceVersion);
+            if (newState == _state)
             {
-                return ImmutableArray.Create<DocumentId>();
+                return this;
             }
 
-            ImmutableArray<DocumentId> documentIds;
-            return _linkedFilesMap.TryGetValue(filePath, out documentIds)
-                ? documentIds
-                : ImmutableArray.Create<DocumentId>();
-        }
-
-        private static ProjectDependencyGraph CreateDependencyGraph(
-            IReadOnlyList<ProjectId> projectIds,
-            ImmutableDictionary<ProjectId, ProjectState> projectStates)
-        {
-            var map = projectStates.Values.Select(state => new KeyValuePair<ProjectId, ImmutableHashSet<ProjectId>>(
-                    state.Id,
-                    state.ProjectReferences.Where(pr => projectStates.ContainsKey(pr.ProjectId)).Select(pr => pr.ProjectId).ToImmutableHashSet()))
-                    .ToImmutableDictionary();
-
-            return new ProjectDependencyGraph(projectIds.ToImmutableArray(), map);
-        }
-
-        private ImmutableDictionary<ProjectId, CompilationTracker> CreateCompilationTrackerMap(ProjectId projectId, ProjectDependencyGraph dependencyGraph)
-        {
-            var builder = ImmutableDictionary.CreateBuilder<ProjectId, CompilationTracker>();
-            var dependencies = dependencyGraph.GetProjectsThatTransitivelyDependOnThisProject(projectId);
-
-            foreach (var projectIdAndTracker in _projectIdToTrackerMap)
-            {
-                var id = projectIdAndTracker.Key;
-                var tracker = projectIdAndTracker.Value;
-
-                if (!tracker.HasCompilation)
-                {
-                    continue;
-                }
-
-                var canReuse = id == projectId || !dependencies.Contains(id);
-                builder.Add(id, canReuse ? tracker : tracker.Fork(tracker.ProjectState));
-            }
-
-            return builder.ToImmutable();
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1896,85 +955,8 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution GetIsolatedSolution()
         {
-            var forkedMap = ImmutableDictionary.CreateRange<ProjectId, CompilationTracker>(
-                _projectIdToTrackerMap.Where(kvp => kvp.Value.HasCompilation)
-                                     .Select(kvp => new KeyValuePair<ProjectId, CompilationTracker>(kvp.Key, kvp.Value.Clone())));
-
-            return this.Branch(projectIdToTrackerMap: forkedMap);
-        }
-
-        // this lock guards all the mutable fields (do not share lock with derived classes)
-        private NonReentrantLock _stateLockBackingField;
-        private NonReentrantLock StateLock
-        {
-            get
-            {
-                return LazyInitializer.EnsureInitialized(ref _stateLockBackingField, NonReentrantLock.Factory);
-            }
-        }
-
-        private WeakReference<Solution> _latestSolutionWithPartialCompilation;
-        private DateTime _timeOfLatestSolutionWithPartialCompilation;
-        private DocumentId _documentIdOfLatestSolutionWithPartialCompilation;
-
-        /// <summary>
-        /// Creates a branch of the solution that has its compilations frozen in whatever state they are in at the time, assuming a background compiler is
-        /// busy building this compilations.
-        /// 
-        /// A compilation for the project containing the specified document id will be guaranteed to exist with at least the syntax tree for the document.
-        /// 
-        /// This not intended to be the public API, use Document.WithFrozenPartialSemantics() instead.
-        /// </summary>
-        internal async Task<Solution> WithFrozenPartialCompilationIncludingSpecificDocumentAsync(DocumentId documentId, CancellationToken cancellationToken)
-        {
-            try
-            {
-                var doc = this.GetDocument(documentId);
-                var tree = await doc.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false);
-
-                using (this.StateLock.DisposableWait(cancellationToken))
-                {
-                    // in progress solutions are disabled for some testing
-                    Workspace ws = this.Workspace as Workspace;
-                    if (ws != null && ws.TestHookPartialSolutionsDisabled)
-                    {
-                        return this;
-                    }
-
-                    Solution currentPartialSolution = null;
-                    if (_latestSolutionWithPartialCompilation != null)
-                    {
-                        _latestSolutionWithPartialCompilation.TryGetTarget(out currentPartialSolution);
-                    }
-
-                    // if we don't have one or it is stale, create a new partial solution
-                    if (currentPartialSolution == null
-                        || (DateTime.UtcNow - _timeOfLatestSolutionWithPartialCompilation).TotalSeconds >= 0.1
-                        || _documentIdOfLatestSolutionWithPartialCompilation != documentId)
-                    {
-                        var tracker = this.GetCompilationTracker(documentId.ProjectId);
-                        var newTracker = tracker.FreezePartialStateWithTree(this, (DocumentState)doc.State, tree, cancellationToken);
-
-                        var newIdToProjectStateMap = _projectIdToProjectStateMap.SetItem(documentId.ProjectId, newTracker.ProjectState);
-                        var newIdToTrackerMap = _projectIdToTrackerMap.SetItem(documentId.ProjectId, newTracker);
-
-                        currentPartialSolution = this.Branch(
-                            idToProjectStateMap: newIdToProjectStateMap,
-                            projectIdToTrackerMap: newIdToTrackerMap,
-                            dependencyGraph: CreateDependencyGraph(_projectIds, newIdToProjectStateMap));
-
-                        _latestSolutionWithPartialCompilation = new WeakReference<Solution>(currentPartialSolution);
-                        _timeOfLatestSolutionWithPartialCompilation = DateTime.UtcNow;
-                        _documentIdOfLatestSolutionWithPartialCompilation = documentId;
-                    }
-
-                    return currentPartialSolution;
-                }
-            }
-            catch (Exception e) when (FatalError.ReportUnlessCanceled(e))
-            {
-                throw ExceptionUtilities.Unreachable;
-            }
+            var newState = _state.GetIsolatedSolution();
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -1982,144 +964,13 @@ namespace Microsoft.CodeAnalysis
         /// </summary>
         public Solution WithDocumentText(IEnumerable<DocumentId> documentIds, SourceText text, PreservationMode mode = PreservationMode.PreserveValue)
         {
-            if (documentIds == null)
+            var newState = _state.WithDocumentText(documentIds, text, mode);
+            if (newState == _state)
             {
-                throw new ArgumentNullException(nameof(documentIds));
+                return this;
             }
 
-            if (text == null)
-            {
-                throw new ArgumentNullException(nameof(text));
-            }
-
-            var solution = this;
-
-            foreach (var documentId in documentIds)
-            {
-                var doc = solution.GetDocument(documentId);
-                if (doc != null)
-                {
-                    SourceText existingText;
-                    if (!doc.TryGetText(out existingText) || existingText != text)
-                    {
-                        solution = solution.WithDocumentText(documentId, text, mode);
-                    }
-                }
-            }
-
-            return solution;
-        }
-
-        internal bool TryGetCompilation(ProjectId projectId, out Compilation compilation)
-        {
-            CheckContainsProject(projectId);
-
-            CompilationTracker tracker;
-            compilation = null;
-
-            return this.TryGetCompilationTracker(projectId, out tracker)
-                && tracker.TryGetCompilation(out compilation);
-        }
-
-        /// <summary>
-        /// Returns the compilation for the specified <see cref="ProjectId"/>.  Can return <code>null</code> when the project
-        /// does not support compilations.
-        /// </summary>
-        internal Task<Compilation> GetCompilationAsync(ProjectId projectId, CancellationToken cancellationToken)
-        {
-            return GetCompilationAsync(GetProject(projectId), cancellationToken);
-        }
-
-        /// <summary>
-        /// Returns the compilation for the specified <see cref="Project"/>.  Can return <code>null</code> when the project
-        /// does not support compilations.
-        /// </summary>
-        internal Task<Compilation> GetCompilationAsync(Project project, CancellationToken cancellationToken)
-        {
-            return project.SupportsCompilation
-                ? this.GetCompilationTracker(project.Id).GetCompilationAsync(this, cancellationToken)
-                : SpecializedTasks.Default<Compilation>();
-        }
-
-        /// <summary>
-        /// Return reference completeness for the given project and all projects this references.
-        /// </summary>
-        internal Task<bool> HasSuccessfullyLoadedAsync(Project project, CancellationToken cancellationToken)
-        {
-            // return HasAllInformation when compilation is not supported. 
-            // regardless whether project support compilation or not, if projectInfo is not complete, we can't gurantee its reference completeness
-            return project.SupportsCompilation
-                ? this.GetCompilationTracker(project.Id).HasSuccessfullyLoadedAsync(this, cancellationToken)
-                : project.Solution.GetProjectState(project.Id).HasAllInformation ? SpecializedTasks.True : SpecializedTasks.False;
-        }
-
-        /// <summary>
-        /// Symbols need to be either <see cref="IAssemblySymbol"/> or <see cref="IModuleSymbol"/>.
-        /// </summary>
-        private static readonly ConditionalWeakTable<ISymbol, ProjectId> s_assemblyOrModuleSymbolToProjectMap =
-            new ConditionalWeakTable<ISymbol, ProjectId>();
-
-        private static void RecordSourceOfAssemblySymbol(ISymbol assemblyOrModuleSymbol, ProjectId projectId)
-        {
-            if (assemblyOrModuleSymbol == null)
-            {
-                return;
-            }
-
-            Contract.ThrowIfNull(projectId);
-
-            // remember which project is associated with this assembly
-            ProjectId tmp;
-            if (!s_assemblyOrModuleSymbolToProjectMap.TryGetValue(assemblyOrModuleSymbol, out tmp))
-            {
-                // use GetValue to avoid race condition exceptions from Add.
-                // the first one to set the value wins.
-                s_assemblyOrModuleSymbolToProjectMap.GetValue(assemblyOrModuleSymbol, _ => projectId);
-            }
-            else
-            {
-                // sanity check: this should always be true, no matter how many times
-                // we attempt to record the association.
-                System.Diagnostics.Debug.Assert(tmp == projectId);
-            }
-        }
-
-        /// <summary>
-        /// Get a metadata reference for the project's compilation
-        /// </summary>
-        internal Task<MetadataReference> GetMetadataReferenceAsync(ProjectReference projectReference, ProjectState fromProject, CancellationToken cancellationToken)
-        {
-            try
-            {
-                // Get the compilation state for this project.  If it's not already created, then this
-                // will create it.  Then force that state to completion and get a metadata reference to it.
-                var tracker = this.GetCompilationTracker(projectReference.ProjectId);
-                return tracker.GetMetadataReferenceAsync(this, fromProject, projectReference, cancellationToken);
-            }
-            catch (Exception e) when (FatalError.ReportUnlessCanceled(e))
-            {
-                throw ExceptionUtilities.Unreachable;
-            }
-        }
-
-        /// <summary>
-        /// Attempt to get the best readily available compilation for the project. It may be a
-        /// partially built compilation.
-        /// </summary>
-        internal MetadataReference GetPartialMetadataReference(
-            ProjectReference projectReference,
-            ProjectState fromProject,
-            CancellationToken cancellationToken)
-        {
-            // Try to get the compilation state for this project.  If it doesn't exist, don't do any
-            // more work.  
-            CompilationTracker state;
-            if (!_projectIdToTrackerMap.TryGetValue(projectReference.ProjectId, out state))
-            {
-                return null;
-            }
-
-            return state.GetPartialMetadataReference(this, fromProject, projectReference, cancellationToken);
+            return new Solution(newState);
         }
 
         /// <summary>
@@ -2136,154 +987,16 @@ namespace Microsoft.CodeAnalysis
             return new SolutionChanges(this, oldSolution);
         }
 
-        internal async Task<bool> ContainsSymbolsWithNameAsync(ProjectId id, Func<string, bool> predicate, SymbolFilter filter, CancellationToken cancellationToken)
-        {
-            var result = GetCompilationTracker(id).ContainsSymbolsWithNameFromDeclarationOnlyCompilation(predicate, filter, cancellationToken);
-            if (result.HasValue)
-            {
-                return result.Value;
-            }
-
-            // it looks like declaration compilation doesn't exist yet. we have to build full compilation
-            var compilation = await GetCompilationAsync(id, cancellationToken).ConfigureAwait(false);
-            if (compilation == null)
-            {
-                // some projects don't support compilations (e.g., TypeScript) so there's nothing to check
-                return false;
-            }
-
-            return compilation.ContainsSymbolsWithName(predicate, filter, cancellationToken);
-        }
-
-        internal async Task<IEnumerable<Document>> GetDocumentsWithNameAsync(ProjectId id, Func<string, bool> predicate, SymbolFilter filter, CancellationToken cancellationToken)
-        {
-            // this will be used to find documents that contain declaration information in IDE cache such as DeclarationSyntaxTreeInfo for "NavigateTo"
-            var trees = GetCompilationTracker(id).GetSyntaxTreesWithNameFromDeclarationOnlyCompilation(predicate, filter, cancellationToken);
-            if (trees != null)
-            {
-                return ConvertTreesToDocuments(id, trees);
-            }
-
-            // it looks like declaration compilation doesn't exist yet. we have to build full compilation
-            var compilation = await GetCompilationAsync(id, cancellationToken).ConfigureAwait(false);
-            if (compilation == null)
-            {
-                // some projects don't support compilations (e.g., TypeScript) so there's nothing to check
-                return SpecializedCollections.EmptyEnumerable<Document>();
-            }
-
-            return ConvertTreesToDocuments(
-                id, compilation.GetSymbolsWithName(predicate, filter, cancellationToken).SelectMany(s => s.DeclaringSyntaxReferences.Select(r => r.SyntaxTree)));
-        }
-
-        private IEnumerable<Document> ConvertTreesToDocuments(ProjectId id, IEnumerable<SyntaxTree> trees)
-        {
-            foreach (var tree in trees)
-            {
-                var document = GetDocument(tree, id);
-                if (document == null)
-                {
-                    // ignore trees that are not known to solution such as VB synthesized trees made by compilation.
-                    continue;
-                }
-
-                yield return document;
-            }
-        }
+        /// <summary>
+        /// Gets the set of <see cref="DocumentId"/>s in this <see cref="Solution"/> with a
+        /// <see cref="TextDocument.FilePath"/> that matches the given file path.
+        /// </summary>
+        public ImmutableArray<DocumentId> GetDocumentIdsWithFilePath(string filePath) => _state.GetDocumentIdsWithFilePath(filePath);
 
         /// <summary>
         /// Gets a <see cref="ProjectDependencyGraph"/> that details the dependencies between projects for this solution.
         /// </summary>
-        public ProjectDependencyGraph GetProjectDependencyGraph()
-        {
-            return _dependencyGraph;
-        }
-
-        private void CheckNotContainsProject(ProjectId projectId)
-        {
-            if (this.ContainsProject(projectId))
-            {
-                throw new InvalidOperationException(WorkspacesResources.The_solution_already_contains_the_specified_project);
-            }
-        }
-
-        private void CheckContainsProject(ProjectId projectId)
-        {
-            if (!this.ContainsProject(projectId))
-            {
-                throw new InvalidOperationException(WorkspacesResources.The_solution_does_not_contain_the_specified_project);
-            }
-        }
-
-        private void CheckNotContainsProjectReference(ProjectId projectId, ProjectReference referencedProject)
-        {
-            if (this.GetProjectState(projectId).ProjectReferences.Contains(referencedProject))
-            {
-                throw new InvalidOperationException(WorkspacesResources.The_project_already_references_the_target_project);
-            }
-        }
-
-        private void CheckNotContainsTransitiveReference(ProjectId fromProjectId, ProjectId toProjectId)
-        {
-            var dependents = _dependencyGraph.GetProjectsThatThisProjectTransitivelyDependsOn(fromProjectId);
-            if (dependents.Contains(toProjectId))
-            {
-                throw new InvalidOperationException(WorkspacesResources.The_project_already_transitively_references_the_target_project);
-            }
-        }
-
-        private void CheckNotSecondSubmissionReference(ProjectId projectId, ProjectId toProjectId)
-        {
-            var projectState = GetProjectState(projectId);
-
-            if (projectState.IsSubmission && GetProjectState(toProjectId).IsSubmission)
-            {
-                if (projectState.ProjectReferences.Any(p => GetProjectState(p.ProjectId).IsSubmission))
-                {
-                    throw new InvalidOperationException(WorkspacesResources.This_submission_already_references_another_submission_project);
-                }
-            }
-        }
-
-        private void CheckNotContainsDocument(DocumentId documentId)
-        {
-            Contract.Requires(!this.ContainsDocument(documentId));
-
-            if (this.ContainsDocument(documentId))
-            {
-                throw new InvalidOperationException(WorkspacesResources.The_solution_already_contains_the_specified_document);
-            }
-        }
-
-        private void CheckNotContainsAdditionalDocument(DocumentId documentId)
-        {
-            Contract.Requires(!this.ContainsAdditionalDocument(documentId));
-
-            if (this.ContainsAdditionalDocument(documentId))
-            {
-                throw new InvalidOperationException(WorkspacesResources.The_solution_already_contains_the_specified_document);
-            }
-        }
-
-        private void CheckContainsDocument(DocumentId documentId)
-        {
-            Contract.Requires(this.ContainsDocument(documentId));
-
-            if (!this.ContainsDocument(documentId))
-            {
-                throw new InvalidOperationException(WorkspacesResources.The_solution_does_not_contain_the_specified_document);
-            }
-        }
-
-        private void CheckContainsAdditionalDocument(DocumentId documentId)
-        {
-            Contract.Requires(this.ContainsAdditionalDocument(documentId));
-
-            if (!this.ContainsAdditionalDocument(documentId))
-            {
-                throw new InvalidOperationException(WorkspacesResources.The_solution_does_not_contain_the_specified_document);
-            }
-        }
+        public ProjectDependencyGraph GetProjectDependencyGraph() => _state.GetProjectDependencyGraph();
 
         /// <summary>
         /// Returns the options that should be applied to this solution. This consists of global options from <see cref="Workspace.Options"/>,
