+++ src/Workspaces/Core/Portable/Workspace/Solution/SolutionState.CompilationTranslationAction.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 
 namespace Microsoft.CodeAnalysis
 {
-    public partial class Solution
+    internal partial class SolutionState
     {
         private abstract partial class CompilationTranslationAction
         {
