+++ src/Workspaces/CoreTest/SolutionTests/SolutionTests.cs
@@ -228,7 +228,7 @@ namespace Microsoft.CodeAnalysis.UnitTests
                 {
                     if (solution.ContainsProject(referenced.ProjectId))
                     {
-                        var referencedMetadata = solution.GetMetadataReferenceAsync(referenced, solution.GetProjectState(project.Id), CancellationToken.None).Result;
+                        var referencedMetadata = solution.State.GetMetadataReferenceAsync(referenced, solution.GetProjectState(project.Id), CancellationToken.None).Result;
                         Assert.NotNull(referencedMetadata);
                         var compilationReference = referencedMetadata as CompilationReference;
                         if (compilationReference != null)
